Display side text
used marrisa's myMsg function to display the result on the side panels.
Used a timeout function so that the current result is displayed on the
side screens after a set period of time ( at the moment 1 minute), this
is when david will acknowledge the result. It gets the current choice by
looking at the contents of the element with the result class for each
different view